(REFACTOR) user/term MVI 적용 및 목업 데이터 분리 (#20)
